Moved scalar field util to separate file
